# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mini Wallet System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/balance:
    get:
      operationId: getBalance
      tags: [wallet]
      summary: Get wallet balance for current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallet/add-money:
    post:
      operationId: addMoney
      tags: [wallet]
      summary: Add money to wallet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMoneyBody"
      responses:
        "200":
          description: Money added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResult"
  /wallet/withdraw:
    post:
      operationId: withdrawMoney
      tags: [wallet]
      summary: Withdraw money from wallet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "200":
          description: Withdrawal successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResult"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
...
    WalletBalance:
      type: object
      properties:
        walletId:
          type: string
        balance:
          type: number
        dailyTransferLimit:
          type: number
        dailyTransferUsed:
          type: number
        currency:
          type: string
      required: [walletId, balance, dailyTransferLimit, dailyTransferUsed, currency]
    AddMoneyBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        note:
          type: string
      required: [amount]
    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        note:
          type: string
      required: [amount]
...
    TransactionResult:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: [INITIATED, PROCESSING, SUCCESS, FAILED, PENDING, REVERSED]
        amount:
          type: number
        message:
          type: string
        idempotencyKey:
          type: string
        isDuplicate:
          type: boolean
      required: [transactionId, status, amount, message]
...